Move test extention to example assets
since dialogic_additions are ignored in git settings, a link will now be used instead. Testers will be adviced to either make a directory link or copy the test extention files to dialogic_additions.

Also added a class name to timeline editor. This is to allow better autocomplete while coding.